requirements.txt
python-telegram-bot==13.7
PyYAML==6.0
